// src/api/sqlmodule/model/submitAttemptResponseReason.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type SubmitAttemptResponseReason = typeof SubmitAttemptResponseReason[keyof typeof SubmitAttemptResponseReason];


export const SubmitAttemptResponseReason = {
  Ok: 'Ok',
  ColumnMismatch: 'ColumnMismatch',
  RowCountMismatch: 'RowCountMismatch',
  ValueMismatch: 'ValueMismatch',
  SqlError: 'SqlError',
  Timeout: 'Timeout',
  NotRun: 'NotRun',
  ResultLimitExceeded: 'ResultLimitExceeded',
} as const;