// src/api/sqlmodule/model/submitAttemptResponseStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type SubmitAttemptResponseStatus = typeof SubmitAttemptResponseStatus[keyof typeof SubmitAttemptResponseStatus];


export const SubmitAttemptResponseStatus = {
  Succeeded: 'Succeeded',
  Error: 'Error',
  TimedOut: 'TimedOut',
} as const;